Added a document which talks about the object tracking algorithms, how to use them, references, etc.
Also took out some un-used functionality from the ObjectTracker class.  Currently, the LINEMOD implementation is there but empty.  I'm not sure what to do about that right now--if they will want to keep it and fill it in later or what, so I'll leave the shell of it for now.

git-svn-id: http://gsoc11-opencv-tracking.googlecode.com/svn/trunk@19 f91c6a58-2d9b-92e0-0c1a-53cd64602a37